src/main/java/com/raexample/pojo/lottoapi/Lotto.java
package com.raexample.pojo.lottoapi;

import java.util.*;

public class Lotto {
    private int lottoId;
    private List<Integer> winningNumbers = null;
    private List<Winner> winners = null;

    public int getLottoId() {
        return lottoId;
    }

    public void setLottoId(int lottoId) {
        this.lottoId = lottoId;
    }

    public List<Integer> getWinningNumbers() {
        return winningNumbers;
    }

    public void setWinningNumbers(List<Integer> winningNumbers) {
        this.winningNumbers = winningNumbers;
    }

    public List<Winner> getWinners() {
        return winners;
    }

    public void setWinners(List<Winner> winners) {
        this.winners = winners;
    }
}
